v2023.06.01 :kiss: Complete New
[instangram] Complete rewrite of instantgram changed the way this software detects images/videos.
No more problems if Instagram changes their frontend!
Added for every post/story a modal to easier select the needed image/video.
Also it is now possible to click on download and it will starts donwloading instead of open a new tab.
In the future, options will be introduced with which one can make personal settings that will be taken into account at the next call.